+++ CHANGES.md
@@ -10,6 +10,7 @@
 -   Autofocus on first element in create form; autofocus on edit element
 -   Add onKeyDown function to Input component #58
 -   Add error to Int field overflow #57
+-   Fix filter and sort button header position #52
 
 ## Version 1.2.0